+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 0.1.2 (2025-07-23)
+* Added Gentoo Linux
+* Added GNOME Terminal to Cinnamon and Budgie packages list (since they don't have a terminal emulator by default)
+* Added GNOME Files to Budgie packages list (since it doesn't have a file manager by default)
+
 ## 0.1.1 (2025-07-19)
 * Added Void Linux
 * Added LightDM GTK greeter to the package list